user-management/pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = ""
authors = ["Abdullah <abdullahasifmf007+github@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.1.1"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
behave = "^1.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
